refactor: alterações para padronização
