+++ src/main/java/br/sc/senac/tcs/controller/ClienteController.java
@@ -1,6 +1,5 @@
 package br.sc.senac.tcs.controller;
 
-import org.hibernate.mapping.List;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.CrossOrigin;
@@ -15,7 +14,6 @@ import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
 import br.sc.senac.tcs.model.entidade.Cliente;
-import br.sc.senac.tcs.model.repository.ClienteRepository;
 import br.sc.senac.tcs.service.ClienteService;
 
 @RestController
