+++ src/main/java/br/sc/senac/tcs/model/entidade/Seguro.java
@@ -3,7 +3,11 @@ package br.sc.senac.tcs.model.entidade;
 import java.time.LocalDate;
 
 import com.fasterxml.jackson.annotation.JsonBackReference;
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+import com.fasterxml.jackson.annotation.JsonIgnoreType;
+import com.fasterxml.jackson.annotation.JsonManagedReference;
 
+import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
@@ -21,20 +25,19 @@ public class Seguro {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private int id;
-//	@OneToOne
-//	@JoinColumn(name = "seguradora_id")
-//	private Seguradora seguradora_id;
+	private Integer id;
+	@OneToOne
+	@JoinColumn(name = "seguradora_id")
+	private Seguradora seguradora;
 	@OneToOne
 	@JoinColumn(name = "veiculo_id")
 	private Veiculo veiculo;
 	@ManyToOne
     @JoinColumn(name = "corretor_id")
 	private Corretor corretor;
-	@ManyToOne
-    @JoinColumn(name = "cliente_id")
-	private Cliente cliente;
 	
+	@Column(name = "cliente_id")
+	private Integer idCliente;
 	private double rcfDanosMateriais;
 	private double rofDanosFisicos;
 	private LocalDate dtInicioVigencia;
