+++ src/main/java/br/sc/senac/tcs/model/repository/CorretorRepository.java
@@ -1,9 +1,17 @@
 package br.sc.senac.tcs.model.repository;
 
-import org.springframework.data.repository.CrudRepository;
+import java.util.List;
+
+import org.springframework.data.jpa.domain.Specification;
+import org.springframework.data.jpa.repository.JpaRepository;
+import org.springframework.stereotype.Repository;
 
 import br.sc.senac.tcs.model.entidade.Corretor;
 
-public interface CorretorRepository extends CrudRepository<Corretor,Integer >{
-    
+@Repository
+public interface CorretorRepository extends JpaRepository<Corretor, Integer> {
+	
+	List<Corretor> findAll(Specification<Corretor> specification);
+	
+	Corretor getReferenceById(Integer id);
 }
