--- src/main/java/br/sc/senac/tcs/repository/Repository.java
@@ -1,6 +1,0 @@
-package br.sc.senac.tcs.repository;
-
-
-public interface Repository {
-
-}
