+++ src/main/java/br/sc/senac/tcs/service/SeguradoraService.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Service;
 
 import br.sc.senac.tcs.exception.CampoInvalidoException;
 import br.sc.senac.tcs.model.entidade.Seguradora;
-import br.sc.senac.tcs.repository.SeguradoraRepository;
+import br.sc.senac.tcs.model.repository.SeguradoraRepository;
 import jakarta.transaction.Transactional;
 
 @Service
